+++ packages/sdk/src/parser/service/idp/schema.ts
@@ -2,6 +2,11 @@ import { z } from "zod";
 
 export const IdPLangSchema = z.enum(["en", "ja"]);
 
+export const IdPUserAuthPolicySchema = z.object({
+  useNonEmailIdentifier: z.boolean().default(false),
+  allowSelfPasswordReset: z.boolean().default(false),
+});
+
 export const IdPSchema = z
   .object({
     name: z.string(),
@@ -12,5 +17,8 @@ export const IdPSchema = z
     ]),
     clients: z.array(z.string()),
     lang: IdPLangSchema.optional(),
+    userAuthPolicy: IdPUserAuthPolicySchema.transform((input) =>
+      IdPUserAuthPolicySchema.parse(input ?? {}),
+    ).optional(),
   })
   .brand("IdPConfig");